tradefed_cluster/requirements.txt
absl-py==0.8.1
antlr3-python-runtime==3.4
bunch==1.0.1
croniter==0.3.30
docker==4.1.0
futures==3.3.0
google-api-python-client==1.7.11
google-auth==1.7.1
google-cloud-storage==1.23.0
httplib2==0.14.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
mock==3.0.5
oauth2client==4.1.3
portpicker==1.3.1
PyHamcrest==1.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
retry==0.9.2
webapp2==2.5.2
WebTest==2.0.33
